+++ inc/flex/fixed_pool.h
@@ -31,7 +31,7 @@ namespace flex
     union
     {
       char mBuffer[N * sizeof(node_type)];
-      uint64_t dummy;
+      long double dummy;
     };
 #endif
   };
