+++ inc/flex/fixed_vector.h
@@ -45,7 +45,7 @@ namespace flex
     union
     {
       char mBuffer[N * sizeof(T)];
-      uint64_t dummy;
+      long double dummy;
     };
 #endif
   };
